+++ Tests/Unit/ComputerManagementDsc.Common.Tests.ps1
@@ -440,6 +440,217 @@ try
                     }
                 }
             }
+
+            Describe 'ComputerManagementDsc.Common\Get-TimeZoneId' {
+                Context '"Get-TimeZone" not available and current timezone is set to "Pacific Standard Time"' {
+                    Mock -CommandName Get-Command -ParameterFilter {
+                        $Name -eq 'Get-TimeZone'
+                    }
+
+                    Mock -CommandName Get-CimInstance -MockWith {
+                        @{
+                            StandardName = 'Pacific Standard Time'
+                        }
+                    }
+
+                    It 'Returns "Pacific Standard Time"' {
+                        Get-TimeZoneId | Should -Be 'Pacific Standard Time'
+                    }
+
+                    It 'Should call expected mocks' {
+                        Assert-MockCalled -CommandName Get-Command -ParameterFilter {
+                            $Name -eq 'Get-TimeZone'
+                        } -Exactly -Times 1
+
+                        Assert-MockCalled -CommandName Get-CimInstance -Exactly -Times 1
+                    }
+                }
+
+                Context '"Get-TimeZone" not available and current timezone is set to "Russia TZ 11 Standard Time"' {
+                    Mock -CommandName Get-Command -ParameterFilter {
+                        $Name -eq 'Get-TimeZone'
+                    }
+
+                    Mock -CommandName Get-CimInstance -MockWith {
+                        @{
+                            StandardName = 'Russia TZ 11 Standard Time'
+                        }
+                    }
+
+                    It 'Returns "Russia Time Zone 11"' {
+                        Get-TimeZoneId | Should -Be 'Russia Time Zone 11'
+                    }
+
+                    It 'Should call expected mocks' {
+                        Assert-MockCalled -CommandName Get-Command -ParameterFilter {
+                            $Name -eq 'Get-TimeZone'
+                        } -Exactly -Times 1
+
+                        Assert-MockCalled -CommandName Get-CimInstance -Exactly -Times 1
+                    }
+                }
+
+                Context '"Get-TimeZone" available and current timezone is set to "Pacific Standard Time"' {
+                    function Get-TimeZone
+                    {
+                    }
+
+                    Mock -CommandName Get-Command -ParameterFilter {
+                        $Name -eq 'Get-TimeZone'
+                    } -MockWith {
+                        'Get-TimeZone'
+                    }
+
+                    Mock -CommandName Get-TimeZone -MockWith {
+                        @{
+                            StandardName = 'Pacific Standard Time'
+                        }
+                    }
+
+                    It 'Returns "Pacific Standard Time"' {
+                        Get-TimeZoneId | Should -Be 'Pacific Standard Time'
+                    }
+
+                    It 'Should call expected mocks' {
+                        Assert-MockCalled -CommandName Get-Command -ParameterFilter {
+                            $Name -eq 'Get-TimeZone'
+                        } -Exactly -Times 1
+
+                        Assert-MockCalled -CommandName Get-TimeZone -Exactly -Times 1
+                    }
+                }
+            }
+
+            Describe 'ComputerManagementDsc.Common\Test-TimezoneId' {
+                Mock Get-TimeZoneId -MockWith {
+                    'Russia Time Zone 11'
+                }
+
+                Context 'current timezone matches desired timezone' {
+                    It 'Should return $true' {
+                        Test-TimezoneId -TimeZoneId 'Russia Time Zone 11' | Should -Be $true
+                    }
+                }
+
+                Context 'current timezone does not match desired timezone' {
+                    It 'Should return $false' {
+                        Test-TimezoneId -TimeZoneId 'GMT Standard Time' | Should -Be $false
+                    }
+                }
+            }
+
+            Describe 'ComputerManagementDsc.Common\Set-TimeZoneId' {
+                Context '"Set-TimeZone" and "Add-Type" is not available, Tzutil Returns 0' {
+                    Mock -CommandName Get-Command -ParameterFilter {
+                        $Name -eq 'Add-Type'
+                    }
+
+                    Mock -CommandName Get-Command -ParameterFilter {
+                        $Name -eq 'Set-TimeZone'
+                    }
+
+                    Mock -CommandName 'TzUtil.exe' -MockWith {
+                        $global:LASTEXITCODE = 0
+                        return 'OK'
+                    }
+
+                    Mock -CommandName Add-Type
+
+                    It 'Should not throw an exception' {
+                        { Set-TimeZoneId -TimezoneId 'Eastern Standard Time' } | Should -Not -Throw
+                    }
+
+                    It 'Should call expected mocks' {
+                        Assert-MockCalled -CommandName Get-Command -ParameterFilter {
+                            $Name -eq 'Add-Type'
+                        } -Exactly -Times 1
+
+                        Assert-MockCalled -CommandName Get-Command -ParameterFilter {
+                            $Name -eq 'Set-TimeZone'
+                        } -Exactly -Times 1
+
+                        Assert-MockCalled -CommandName TzUtil.exe -Exactly -Times 1
+                        Assert-MockCalled -CommandName Add-Type -Exactly -Times 0
+                    }
+                }
+
+                Context '"Set-TimeZone" is not available but "Add-Type" is available' {
+                    Mock -CommandName Get-Command -ParameterFilter {
+                        $Name -eq 'Add-Type'
+                    } -MockWith {
+                        'Add-Type'
+                    }
+
+                    Mock -CommandName Get-Command -ParameterFilter {
+                        $Name -eq 'Set-TimeZone'
+                    }
+
+                    Mock -CommandName 'TzUtil.exe' -MockWith {
+                        $global:LASTEXITCODE = 0
+                        return 'OK'
+                    }
+
+                    Mock -CommandName Add-Type
+                    Mock -CommandName Set-TimeZoneUsingDotNet
+
+                    It 'Should not throw an exception' {
+                        { Set-TimeZoneId -TimezoneId 'Eastern Standard Time' } | Should -Not -Throw
+                    }
+
+                    It 'Should call expected mocks' {
+                        Assert-MockCalled -CommandName Get-Command -ParameterFilter {
+                            $Name -eq 'Add-Type'
+                        } -Exactly -Times 1
+
+                        Assert-MockCalled -CommandName Get-Command -ParameterFilter {
+                            $Name -eq 'Set-TimeZone'
+                        } -Exactly -Times 1
+
+                        Assert-MockCalled -CommandName TzUtil.exe -Exactly -Times 0
+                        Assert-MockCalled -CommandName Add-Type -Exactly -Times 0
+                        Assert-MockCalled -CommandName Set-TimeZoneUsingDotNet -Exactly -Times 1
+                    }
+                }
+
+                Context '"Set-TimeZone" is available' {
+                    function Set-TimeZone
+                    {
+                        param
+                        (
+                            [System.String]
+                            $id
+                        )
+                    }
+
+                    Mock -CommandName Get-Command -ParameterFilter {
+                        $Name -eq 'Add-Type'
+                    }
+
+                    Mock -CommandName Get-Command -ParameterFilter {
+                        $Name -eq 'Set-TimeZone'
+                    } -MockWith {
+                        'Set-TimeZone'
+                    }
+
+                    Mock -CommandName Set-TimeZone
+
+                    It 'Should not throw an exception' {
+                        { Set-TimeZoneId -TimezoneId 'Eastern Standard Time' } | Should -Not -Throw
+                    }
+
+                    It 'Should call expected mocks' {
+                        Assert-MockCalled -CommandName Get-Command -ParameterFilter {
+                            $Name -eq 'Add-Type'
+                        } -Exactly -Times 0
+
+                        Assert-MockCalled -CommandName Get-Command -ParameterFilter {
+                            $Name -eq 'Set-TimeZone'
+                        } -Exactly -Times 1
+
+                        Assert-MockCalled -CommandName Set-TimeZone -Exactly -Times 1
+                    }
+                }
+            }
         }
     }
 }